WorkerFactoryをカスタマイズできるようにする
仮コミット
