+++ iplass-core/src/main/java/org/iplass/mtp/impl/async/rdb/workers/WorkerCallable.java
@@ -99,7 +99,7 @@ public class WorkerCallable implements Callable<Void> {
 							} else {
 								mtpLogger.warn("queue:" + queue.getConfig().getName() + "'s task:" + task.getTaskId() + "(tenantId:" + task.getTenantId() + ") is failed(cause:" + e + "). re-run after a while.");
 							}
-							break;
+							throw e;
 						case ABORT_LOG_FATAL:
 						case ABORT:
 							Transaction.required(transaction -> {
@@ -122,9 +122,9 @@ public class WorkerCallable implements Callable<Void> {
 									logger.warn("task status cant update to [abort].may be another process update status." + task);
 								}
 							});
-							break;
+							throw e;
 						default:
-							break;
+							throw e;
 						}
 					} finally {
 						ec.removeAttribute(AsyncTaskContextImpl.EXE_CONTEXT_ATTR_NAME);
